# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gerenciamento de fazendas leiteiras\nSistema completo para gestão de animais, coletas de leite, reproduções, vendas e dívidas",
        "title": "FazendaPro API",
        "contact": {
            "name": "Suporte FazendaPro",
            "email": "suporte@fazendapro.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados de um usuário específico pelo email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obter usuário por email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email do usuário",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um novo usuário no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Criar usuário",
                "parameters": [
                    {
                        "description": "Dados do novo usuário",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateUserRequest": {
            "description": "Dados necessários para criar um novo usuário",
            "type": "object",
            "properties": {
                "person": {
                    "description": "Dados da pessoa",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Person"
                        }
                    ]
                },
                "user": {
                    "description": "Dados do usuário",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Animal": {
            "type": "object",
            "properties": {
                "animalName": {
                    "type": "string"
                },
                "animalType": {
                    "type": "integer"
                },
                "birthDate": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "castrated": {
                    "type": "boolean"
                },
                "confinement": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentBatch": {
                    "type": "integer"
                },
                "earTagNumberLocal": {
                    "type": "integer"
                },
                "earTagNumberRegister": {
                    "type": "integer"
                },
                "farm": {
                    "$ref": "#/definitions/models.Farm"
                },
                "farmID": {
                    "type": "integer"
                },
                "father": {
                    "$ref": "#/definitions/models.Animal"
                },
                "fatherID": {
                    "type": "integer"
                },
                "fertilization": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "mother": {
                    "$ref": "#/definitions/models.Animal"
                },
                "motherID": {
                    "type": "integer"
                },
                "photo": {
                    "type": "string"
                },
                "purpose": {
                    "type": "integer"
                },
                "sex": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Company": {
            "type": "object",
            "properties": {
                "companyName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "farmCNPJ": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Expense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "farm": {
                    "$ref": "#/definitions/models.Farm"
                },
                "farmID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Farm": {
            "type": "object",
            "properties": {
                "animals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Animal"
                    }
                },
                "company": {
                    "$ref": "#/definitions/models.Company"
                },
                "companyID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Expense"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.Person": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_access": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "farm": {
                    "$ref": "#/definitions/models.Farm"
                },
                "farm_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "person": {
                    "$ref": "#/definitions/models.Person"
                },
                "person_id": {
                    "type": "integer"
                },
                "user_farms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserFarm"
                    }
                }
            }
        },
        "models.UserFarm": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "farm": {
                    "$ref": "#/definitions/models.Farm"
                },
                "farm_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e depois o token JWT. Exemplo: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}